move thread-locals to separate module
